--- docs/.gitignore
@@ -1,11 +1,0 @@
-/.quarto/
-reference/*.qmd
-reference/_sidebar.yml
-_site
-_inv
-*.quarto_ipynb
-*_files/execute-results
-*_files/figure-html
-objects.json
-_sidebar.yml
-_styles-quartodoc.css
