--- docs/_quarto.yml
@@ -1,70 +1,0 @@
-project:
-  type: website
-
-metadata-files:
-  - reference/_sidebar.yml
-
-filters:
-  - interlinks
-
-interlinks:
-  sources:
-    langchain:
-      url: https://python.langchain.com/api_reference/
-    networkx:
-      url: https://networkx.org/documentation/stable/
-    numpy:
-      url: https://numpy.org/doc/stable/
-    python:
-      url: https://docs.python.org/3/
-
-website:
-  title: Graph RAG
-  twitter-card: true
-  google-analytics: "G-7DS0D7NZML"
-  navbar:
-    background: primary
-    search: true
-    left:
-      - file: get-started/overview.qmd
-        text: Get Started
-      - examples/index.qmd
-      - posts/index.qmd
-      - reference/index.qmd
-    right:
-      - icon: github
-        href: https://github.com/datastax/graph-rag
-      - icon: rss
-        href: index.xml
-  sidebar:
-    - id: get-started
-      title: Get Started
-      style: floating
-      align: left
-      contents:
-        - get-started/overview.qmd
-  page-footer:
-    border: true
-    left: "© Copyright 2025, Graph RAG Contributors."
-  repo-url: https://github.com/datastax/graph-rag/
-  repo-subdir: docs
-  repo-actions:
-    - edit
-    - source
-  site-url: https://datastax.github.io/graph-rag
-
-format:
-  html:
-    theme: pulse
-    toc: true
-    anchor-sections: true
-    css:
-      - css/_styles-quartodoc.css
-      - css/styles.css
-
-    link-external-icon: true
-    link-external-newwindow: true
-    link-external-filter: '^(?:http:|https:)\/\/datastax.github.io\/graph-rag'
-
-
-
