--- docs/_quartodoc.yml
@@ -1,148 +1,0 @@
-quartodoc:
-  package: null
-  parser: numpy
-  style: pkgdown
-  render_interlinks: true
-  renderer:
-    style: markdown
-    table_style: description-list
-  options:
-    include_attributes: true
-    signature_name: full
-  title: 'API Reference'
-  sidebar:
-    file: "reference/_sidebar.yml"
-    style: docked
-    search: true
-    collapse-level: 4
-    id: "reference"
-    contents:
-      - "{{ contents }}"
-  sections:
-
-  - title: graph_retriever
-    desc: |
-      Base package providing traversal functionality.
-    package: graph_retriever
-    contents:
-    - kind: page
-      path: traversal
-      summary:
-        name: Traversal
-        desc: "Methods to perform Graph Traversal"
-      flatten: true
-      contents:
-      - atraverse
-      - traverse
-    - name: Adapter
-      package: graph_retriever.adapters
-    - name: Content
-    - kind: page
-      path: node_types
-      summary:
-        name: Node types
-        desc: "Types related to working with Nodes"
-      flatten: true
-      contents:
-      - Node
-      - Edge
-      - MetadataEdge
-      - IdEdge
-      - Edges
-      - EdgeFunction
-    - kind: page
-      path: edge_types
-      summary:
-        name: Edge types
-        desc: "Types related to defining Edges"
-      flatten: true
-      contents:
-      - EdgeSpec
-      - Id
-      - name: MetadataEdgeFunction
-        package: 'graph_retriever.edges:metadata'
-  - subtitle: 'strategies'
-    package: 'graph_retriever.strategies'
-    contents:
-    - name: Eager
-    - name: Mmr
-    - name: Scored
-    - name: Strategy
-
-  - title: langchain_graph_retriever
-    desc: |
-      Traversal integrated as a LangChain Retriever supporting various VectorStores.
-    package: langchain_graph_retriever
-    contents:
-    - kind: page
-      path: GraphRetriever
-      summary:
-        name: GraphRetriever
-        desc: "Provides a graph-based retriever combining vector search and graph traversal"
-      flatten: true
-      contents:
-      - name: GraphRetriever
-        exclude:
-        - apply_extra
-      - GraphRetriever._aget_relevant_documents
-      - GraphRetriever._get_relevant_documents
-    - name: document_graph
-  - subtitle: 'document_transformers'
-    contents:
-    - name: GLiNEREntityExtractor
-      package: 'langchain_graph_retriever.document_transformers:gliner'
-    - name: HierarchyLinkExtractor
-      package: 'langchain_graph_retriever.document_transformers:hierarchy'
-    - name: HtmlHyperlinkExtractor
-      package: 'langchain_graph_retriever.document_transformers:html_hyperlink'
-    - name: KeybertKeywordExtractor
-      package: 'langchain_graph_retriever.document_transformers:keybert'
-    - name: MetadataDenormalizer
-      package: 'langchain_graph_retriever.document_transformers:metadata_denormalizer'
-    - name: SpacyNERTransformer
-      package: 'langchain_graph_retriever.document_transformers:spacy'
-  - subtitle: 'adapters'
-    contents:
-    - name: AstraAdapter
-      package: 'langchain_graph_retriever.adapters:astra'
-    - name: CassandraAdapter
-      package: 'langchain_graph_retriever.adapters:cassandra'
-    - name: ChromaAdapter
-      package: 'langchain_graph_retriever.adapters:chroma'
-    - name: InMemoryAdapter
-      package: 'langchain_graph_retriever.adapters:in_memory'
-    - name: OpenSearchAdapter
-      package: 'langchain_graph_retriever.adapters:open_search'
-    - name: LangchainAdapter
-      package: 'langchain_graph_retriever.adapters:langchain'
-    - name: DenormalizedAdapter
-      package: 'langchain_graph_retriever.adapters:langchain'
-    - name: 'inference'
-      package: 'langchain_graph_retriever.adapters'
-      members:
-        - infer_adapter
-
-  - title: graph_rag_example_helpers
-    desc: |
-      Helpers used in various examples and notebooks.
-    package: graph_rag_example_helpers
-    contents:
-    - kind: page
-      path: environment
-      summary:
-        name: Environment
-        desc: "Helpers to setup the environment"
-      flatten: true
-      contents:
-      - initialize_environment
-      - Environment
-  - subtitle: datasets
-    package: 'graph_rag_example_helpers.datasets'
-    contents:
-    - name: animals
-      members:
-        - fetch_documents
-    - name: wikimultihop
-      members:
-        - aload_2wikimultihop
-        - BatchPreparer
