--- docs/css/styles.css
@@ -1,64 +1,0 @@
-@import url('https://rsms.me/inter/inter.css');
-html { font-family: 'Inter', sans-serif; }
-@supports (font-variation-settings: normal) {
-  html { font-family: 'Inter var', sans-serif; }
-}
-
-:root {
-  --bs-body-font-family: inter;
-  --bs-body-font-size: 16px;
-}
-
-/* css styles */
-
-.cell-output pre code {
-  white-space: pre-wrap;
-}
-
-
-/* sidebar formatting */
-
-.sidebar a.nav-link {
-font-size: 14.4px;
-font-weight: 400;
-}
-
-.sidebar-item-container .text-start {
-font-weight: 600;
-font-size: 14.4px !important;
-}
-
-.sidebar-item-text {
-font-size: 14px;
-}
-
-.sidebar-item {
-margin-top: 0px;
-}
-
-.sidebar-item-section {
-padding-top: 22px;
-}
-
-.sidebar-section {
-  padding-left: 0px !important;
-}
-
-.sidebar-item-section .sidebar-item-section {
-padding-top: 8px;
-padding-left: 10px;
-}
-
-
-td:first-child {
-text-wrap: nowrap;
-text-size-adjust: 100%;
-}
-
-td:first-child a {
-word-break: normal;
-}
-
-.sidebar-menu-container > ul > li:first-child .menu-text {
-  font-size: 16px;
-}
