+++ docs/examples/.gitignore
@@ -1,2 +1,2 @@
+*.jrnl
 para_with_hyperlink.zip
-load_2wikimultihop.jrnl
