--- docs/posts/index.qmd
@@ -1,8 +1,0 @@
----
-title: "Blog"
-listing:
-  sort: "date desc"
-  type: default
-  categories: true
-  feed: true
----
