+++ docs/reference/graph_rag_example_helpers/index.md
@@ -1,0 +1,6 @@
+# graph_rag_example_helpers
+
+::: graph_rag_example_helpers
+    options:
+        show_submodules: true
+        show_root_toc_entry: false
