+++ docs/reference/graph_retriever/adapters.md
@@ -1,0 +1,5 @@
+# graph_retriever.adapters
+
+::: graph_retriever.adapters
+    options:
+      show_root_toc_entry: false
