+++ docs/reference/graph_retriever/edges.md
@@ -1,0 +1,5 @@
+# graph_retriever.edges
+
+::: graph_retriever.edges
+ options:
+   show_root_toc_entry: false
