+++ docs/reference/graph_retriever/index.md
@@ -1,0 +1,10 @@
+# graph_retriever
+
+::: graph_retriever
+    options:
+      show_root_toc_entry: false
+      members:
+        - Content
+        - Node
+        - atraverse
+        - traverse
