+++ docs/reference/graph_retriever/strategies.md
@@ -1,0 +1,5 @@
+# graph_retriever.strategies
+
+::: graph_retriever.strategies
+    options:
+      show_root_toc_entry: false
