+++ docs/reference/graph_retriever/utils.md
@@ -1,0 +1,5 @@
+# graph_retriever.utils
+
+::: graph_retriever.utils
+ options:
+   show_root_toc_entry: false
