+++ docs/reference/index.md
@@ -1,0 +1,4 @@
+# API Reference
+
+* [`graph-retriever`](./graph_retriever/index.md) contains the core graph traversal logic.
+* [`langchain-graph-retriever`](./langchain_graph_retriever/index.md) contains a [`GraphRetriever`][langchain_graph_retriever.GraphRetriever] and store adapters for use with LangChain.
