+++ docs/reference/langchain_graph_retriever/adapters.md
@@ -1,0 +1,6 @@
+# langchain_graph_retriever.adapters
+
+::: langchain_graph_retriever.adapters
+    options:
+      show_root_toc_entry: false
+      show_submodules: true
