+++ docs/reference/langchain_graph_retriever/index.md
@@ -1,0 +1,5 @@
+# langchain_graph_retriever
+
+::: langchain_graph_retriever
+    options:
+      show_root_toc_entry: false
