+++ packages/graph-retriever/src/graph_retriever/adapters/base.py
@@ -6,7 +6,7 @@ from collections.abc import Iterable, Sequence
 from typing import Any
 
 from graph_retriever.content import Content
-from graph_retriever.types import Edge, IdEdge, MetadataEdge
+from graph_retriever.edges import Edge, IdEdge, MetadataEdge
 from graph_retriever.utils.run_in_executor import run_in_executor
 from graph_retriever.utils.top_k import top_k
