+++ packages/graph-retriever/src/graph_retriever/strategies/scored.py
@@ -1,6 +1,8 @@
 import heapq
 from collections.abc import Callable, Iterable
+from typing import Annotated
 
+from pydantic import WithJsonSchema
 from typing_extensions import override
 
 from graph_retriever.strategies.base import Strategy
@@ -19,7 +21,10 @@ class _ScoredNode:
 class Scored(Strategy):
     """Strategy selecing nodes using a scoring function."""
 
-    scorer: Callable[[Node], float]
+    scorer: Annotated[
+        Callable[[Node], float],
+        WithJsonSchema({"type": "callable", "examples": [1, 0, -1]}),
+    ]
     per_iteration_limit: int | None = None
 
     _nodes: list[_ScoredNode] = []
