+++ packages/langchain-graph-retriever/tests/adapters/test_cassandra.py
@@ -3,7 +3,7 @@ from collections.abc import Iterator
 from typing import Union
 
 import pytest
-from graph_retriever import Adapter
+from graph_retriever.adapters import Adapter
 from graph_retriever.testing.adapter_tests import AdapterComplianceSuite
 from langchain_core.documents import Document
 from langchain_core.embeddings import Embeddings
