+++ packages/langchain-graph-retriever/tests/adapters/test_in_memory.py
@@ -1,5 +1,5 @@
 import pytest
-from graph_retriever import Adapter
+from graph_retriever.adapters import Adapter
 from graph_retriever.testing.adapter_tests import AdapterComplianceSuite
 from langchain_core.documents import Document
 from langchain_core.embeddings import Embeddings
