+++ packages/langchain-graph-retriever/tests/test_document_graph.py
@@ -2,7 +2,7 @@ from typing import cast
 
 import networkx as nx
 import pytest
-from graph_retriever import Id
+from graph_retriever.edges import Id
 from langchain_core.documents import Document
 from langchain_graph_retriever.document_graph import create_graph, group_by_community
